+++ src/containers/Home/GetStarted/GetStarted.scss
@@ -5,7 +5,7 @@
   align-items: center;
   display: flex;
   flex-direction: column;
-  margin-bottom: 120px;
+  margin-bottom: 84px;
   padding: 0 72px;
 
   @media (max-width: 768px) {
