+++ src/containers/Home/GetStarted/GetStartedCard/GetStartedCard.scss
@@ -40,7 +40,7 @@ $padding-768: 24px;
     align-items: center;
     display: flex;
     flex-direction: row;
-    margin-bottom: 35px;
+    margin-bottom: 27px;
   }
 
   &__details-button {
