+++ src/containers/Home/HomeFaq/FaqDropdownCard/FaqDropdownCard.scss
@@ -1,0 +1,47 @@
+@import 'styles/font-mixins';
+
+.FaqDropdownCard {
+  background-color: var(--color-white);
+  border-radius: 8px;
+  box-shadow: 0px 0px 24px 0px rgba(229, 234, 244, 0.58);
+  margin: 32px 40px;
+  min-height: 80px;
+  padding: 24px;
+
+  &__dropdown-container {
+    align-items: center;
+    display: flex;
+    justify-content: space-between;
+  }
+
+  &__answer {
+    @extend %h3;
+    color: var(--color-sail-gray-500);
+    display: none;
+    font-weight: var(--font-weight-regular);
+
+    &--expanded {
+      display: flex;
+      flex-direction: column;
+      margin-top: 30px;
+    }
+  }
+
+  &__left-container {
+    align-items: center;
+    display: flex;
+
+    @media (max-width: 700px) {
+      align-items: flex-start;
+      flex-direction: column;
+      width: 100%;
+    }
+  }
+
+  &__question {
+    color: var(--color-sail-gray-700);
+    font-weight: var(--font-weight-medium);
+    margin-bottom: 8px;
+    @extend %h3;
+  }
+}
