+++ src/containers/Home/HomeFaq/HomeFaq.scss
@@ -1,0 +1,34 @@
+@import 'styles/font-mixins';
+
+.HomeFaq {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  margin: 0 auto;
+  margin-bottom: 130px;
+  max-width: 1366px;
+
+  @media (max-width: 414px) {
+    margin-bottom: 106px;
+  }
+
+  &__title {
+    @extend %display2;
+    font-weight: var(--font-weight-semibold);
+    margin-bottom: 72px;
+    text-align: center;
+
+    @media (max-width: 414px) {
+      margin-bottom: 64px;
+    }
+  }
+
+  &__main {
+    margin-bottom: 82px;
+    width: 100%;
+
+    @media (max-width: 414px) {
+      margin-bottom: 74px;
+    }
+  }
+}
